Rename Pages file and add accessibility role
